Lili and the lost balloon - Final project CS50 - Harvard
Hello!
My name is Ricardo and I’m from São Paulo, Brazil.

I created an entire level of a 2D game.
The name is Lili and the lost balloon, inspired by my daughter, Alice.
This is a classic platform adventure, where your objective is finding the lost balloon.

The game was posted to be played directly by browser. The link is:
https://ricardocadete.itch.io/lili-and-the-lost-balloon

I utilized the Godot platform to make the game.

The architecture is built around the concept of a tree of "nodes". Nodes are organized into "scenes", which are groups of reusable, unstable, inheritable, and nested nodes. All game resources, including scripts and graphical assets, are saved as part of the computer's file system (rather than in a database).

Types of Programming Languages I used at Godot:
GDScript: The main language of the program, a language that has a syntax very similar to Python, but optimized for creating games.
C#: Added in version 3.0.
GDNative: Similar to the C++ language, used to modify the engine itself.

It was a very big challenge, but it served to prove and improve many things I learned during the course.
Being able to make use of classes, global variables and "OOP".
In particular, I really enjoyed implementing the physics of the game, I constantly needed to revise speed and gravity so that the jump didn't get too high and you could access an area very easily and things like that.

I also tried to vary the enemies a little, not only the "sprite" but the way they attack you, playing magic, following you to attack or simply giving you damage when touching.
I put in some interesting mechanics like block push and trampoline.

Implementing shaders was a very interesting thing to do. Shaders are unique programs that run on the GPU. They are used to specify how to take mesh data (vertex positions, colors, normals, etc.) and draw it on the canvas. I used this feature for the dust on Lili's feet as she runs and for the confetti that the balloon releases when the level is over.

In Godot, you can animate anything available in the Inspector, such as Node transforms, sprites, UI elements, particles, visibility and color of materials, and so on. You can also modify values of script variables and call any function.

Getting the level design work done was really cool, creating the terrain by scattering moving platforms and moving the background to make it all more dynamic.

Finally, I was a little worried that I couldn't show the game to people in an easy way and that they could also play without having to install the platform on their computer. Luckily I was able to export the game to a web version the HTML5 export allows games made with Godot Engine to be published to browsers. User browser must support WebAssembly and WebGL.

I had a lot of fun making this environment, implementing ideas and fixing bug’s and I hope you can test and enjoying the experience.

Thank you for everyone and this was my CS50 Final Project.